proxmox/lbprox/requirements.txt
click
jinja2
proxmoxer
prettytable
requests
PyYAML
ipdb
paramiko
